First Push
AR:Dueling

This is the project that I have been working on, on and off for the last year or so. I
recently updated the main charather model and that broke alot of animation things so the
game is not currently playable. However, I thought you guys might still get value out of
looking through the scripts so I decided to include it.

Thank you !

(project was made using Unity 2018.2.20f1)